requirements.txt
aiofiles==0.4.0
aniso8601==6.0.0
attrs==19.1.0
blinker==1.4
certifi==2019.3.9
Click==7.0
falcon==2.0.0
Flask==1.0.2
flask-buzz==0.1.14
Flask-Cors==3.0.7
flask-praetorian==0.5.3
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
h11==0.8.1
h2==3.1.0
hpack==3.0.0
Hypercorn==0.6.0
hyperframe==5.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
multidict==4.5.2
passlib==1.7.1
pendulum==2.0.4
psycopg2-binary==2.8.2
py-buzz==0.3.7
PyJWT==1.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
python-dotenv==0.10.2
pytoml==0.1.20
pytz==2019.1
pytzdata==2019.1
schedule==0.6.0
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.3
typing-extensions==3.7.2
urllib3==1.25.2
uWSGI==2.0.18
Werkzeug==0.15.2
wsproto==0.14.0
